Added graceful command failures
Reorganized command handling so that commands return their
messages and errors, which cleans up the replying to messages
a lot in index.js.  Also, added new error handling with try/catch
so when something fails we can return an appropriate error.

Fix start-prod issue in package.json

Removed unnecessary console.logs

+++ commands/verify.js
@@ -8,15 +8,36 @@ const addVerifiedUser = async (user_id, email, team) => {
 
 const verify = async (code, user_id, client) => {
     const doc = {user_id, verification_code: code};
-    const userVerification = await verification.findOne(doc);
-    if (!userVerification || userVerification.verification_code !== code) {
-        return 'incorrect code';
+    let userVerification = null;
+    try {
+        userVerification = await verification.findOne(doc);
     }
+    catch (err) {
+        console.log('Caught exception in /verify', err);
+        return ['Error when calling database', null];
+    }
+
+    if (!userVerification || userVerification.verification_code !== code) {
+        return [null, 'The code you entered is not correct. Please try again.'];
+    }
+
     const email = userVerification.email;
     const role = await updateUserRoles(email, user_id, client);
-    addVerifiedUser(user_id, email, role);
-    await verification.deleteOne(doc);
-    return 'success';
+    if (!role) {
+        // user not found in sheet
+        return [null, 'You don\'t appear to be listed in the member directory.  If this is an error, please contact leadership.']
+    }
+
+    try {
+        addVerifiedUser(user_id, email, role);
+        await verification.deleteOne(doc);
+    }
+    catch (err) {
+        console.log('Caught exception in /verify', err);
+        return ['Error when updating user roles.', null];
+    }
+
+    return [null, 'Successfully verified your account.'];
 };
 
 module.exports = verify;
